requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-structlog
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
django==5.0.3
    # via
    #   -r requirements.in
    #   django-structlog
django-admin==2.0.2
    # via -r requirements.in
django-excel-response2==3.0.6
    # via django-admin
django-ipware==6.0.4
    # via django-structlog
django-six==1.0.5
    # via
    #   django-admin
    #   django-excel-response2
django-structlog==8.0.0
    # via -r requirements.in
excel-base==1.0.4
    # via django-excel-response2
googleapis-common-protos==1.63.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.62.1
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.6
    # via requests
importlib-metadata==7.0.0
    # via opentelemetry-api
isoweek==1.3.3
    # via timeconvert
opentelemetry-api==1.24.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-distro==0.45b0
    # via -r requirements.in
opentelemetry-exporter-otlp==1.24.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.24.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.24.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.24.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.45b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-django==0.45b0
    # via -r requirements.in
opentelemetry-instrumentation-urllib3==0.45b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.45b0
    # via opentelemetry-instrumentation-django
opentelemetry-proto==1.24.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.24.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.45b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.45b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
protobuf==4.25.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
python-dateutil==2.9.0.post0
    # via timeconvert
python-ipware==2.0.2
    # via django-ipware
requests==2.31.0
    # via opentelemetry-exporter-otlp-proto-http
screen==1.0.1
    # via excel-base
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
structlog==24.1.0
    # via
    #   -r requirements.in
    #   django-structlog
timeconvert==3.0.13
    # via excel-base
typing-extensions==4.10.0
    # via opentelemetry-sdk
tzlocal==5.2
    # via timeconvert
urllib3==2.2.1
    # via requests
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-urllib3
xlwt==1.3.0
    # via excel-base
zipp==3.18.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
